+++ Nos/Service/FileStorage/FileStorageAPIClient.swift
@@ -165,7 +165,7 @@ class NostrBuildAPIClient: FileStorageAPIClient {
         }
         var jsonEvent = JSONEvent(
             pubKey: keyPair.publicKeyHex,
-            kind: .auth,
+            kind: .httpAuth,
             tags: tags,
             content: ""
         )