+++ Nos/Service/Relay/PagedRelaySubscription.swift
@@ -5,6 +5,12 @@ import Combine
 /// This class manages a Filter and fetches events in reverse-chronological order as `loadMore()` is called. This
 /// can be used to paginate a list of events. The underlying relay subscriptions will be deallocated when this object
 /// goes out of scope.  
+///
+/// Paging in Nostr is very different from traditional HTTP paging, because we can't just ask for "the next 20 events
+/// after index 100". Instead we have to use dates and ask for "the next 20 events older than X". Moreover because we
+/// are fetching from a lot of relays the date X is different for every relay. `PagedRelaySubscription` abstracts away
+/// these details, so the caller basically only needs to know what kinds of events they want from what relays, and then
+/// call `loadMore()` whenever the user scrolls a page.
 @RelaySubscriptionManagerActor
 class PagedRelaySubscription {
     let startDate: Date
@@ -69,9 +75,10 @@ class PagedRelaySubscription {
         }
     }
     
-    /// Instructs the pager to load older events for the given `filter` by decrementing the `until` parameter on the 
-    /// `Filter` and updating all its managed subscriptions.
-    func loadMore() {
+    /// Instructs the pager to load the next page of events from each relay. The given date should be roughly the date
+    /// of the content the user is looking at and is used to put a given realy into "catch up" mode where we fetch
+    /// more events to catch up what the user is looking at.
+    func loadMore(displayingContentAt displayedDate: Date? = nil) {
         Task { [self] in
             // Remove old subscriptions
             for subscriptionID in pagedSubscriptionIDs {
@@ -82,17 +89,38 @@ class PagedRelaySubscription {
             
             // Open new subscriptions
             for relayAddress in relayAddresses {
-                let newPageStartDate = oldestEventByRelay[relayAddress] ?? startDate
-                var newPageFilter = self.filter
-                newPageFilter.until = newPageStartDate
-                newPageFilter.keepSubscriptionOpen = false
                 
-                let pagedEventSubscription = await subscriptionManager.queueSubscription(
-                    with: newPageFilter, 
+                // To fetch the next "page" we need to know what the last event we got was, then we ask for the next
+                // `limit` events older than that.
+                let nextPageStartDate = oldestEventByRelay[relayAddress] ?? startDate
+                var nextPageFilter = self.filter
+                nextPageFilter.until = nextPageStartDate
+                nextPageFilter.keepSubscriptionOpen = false
+                
+                // If the most recent event we got is older than what the user is looking at, open an extra subscription
+                // with no limit so we can "catch up".
+                if let displayedDate, nextPageStartDate >= displayedDate {
+                    var catchUpFilter = nextPageFilter
+                    catchUpFilter.since = displayedDate
+                    catchUpFilter.until = nextPageStartDate
+                    catchUpFilter.limit = nil
+                    let catchUpSubscription = await subscriptionManager.queueSubscription(
+                        with: nextPageFilter, 
+                        to: relayAddress
+                    )
+                    pagedSubscriptionIDs.insert(catchUpSubscription.id)
+                    
+                    nextPageFilter.until = displayedDate
+                }
+                
+                let nextPageSubscription = await subscriptionManager.queueSubscription(
+                    with: nextPageFilter, 
                     to: relayAddress
                 )
                 
-                pagedEventSubscription.events
+                /// Keep track of the oldest event seen for this relay so we can use it when it's time to load the next
+                /// page.
+                nextPageSubscription.events
                     .sink { [weak self] jsonEvent in
                         Task {
                             await self?.track(event: jsonEvent, from: relayAddress)
@@ -100,17 +128,21 @@ class PagedRelaySubscription {
                     }
                     .store(in: &cancellables)
                 
-                pagedSubscriptionIDs.insert(pagedEventSubscription.id)
+                pagedSubscriptionIDs.insert(nextPageSubscription.id)
                 
                 await subscriptionManager.processSubscriptionQueue()
             }
         }
     }
     
+    /// Used to record the oldest event we've seen from a relay. We need this because `track(event:from:)` is 
+    /// nonisolated so it can be called from a Combine chain.
     func updateOldestEvent(for relay: URL, to date: Date) {
         oldestEventByRelay[relay] = date
     }
     
+    /// Records the `created_at` date from given event if it's the oldest one we've seen so far. This information
+    /// is needed to load the next page when it's time. 
     nonisolated func track(event: JSONEvent, from relay: URL) async {
         if let oldestSeen = await oldestEventByRelay[relay] {
             if event.createdDate < oldestSeen {