+++ Nos/Views/Home/HomeFeedView.swift
@@ -11,72 +11,60 @@ struct HomeFeedView: View {
     @Environment(CurrentUser.self) var currentUser
     @ObservationIgnored @Dependency(\.analytics) private var analytics
 
-    @FetchRequest private var authors: FetchedResults<Author>
-    
     @State private var lastRefreshDate = Date(
-        timeIntervalSince1970: Date.now.timeIntervalSince1970 + Double(Self.initialLoadTime)
+        timeIntervalSince1970: Date.now.timeIntervalSince1970 + Double(Self.staticLoadTime)
     )
     @State private var isVisible = false
     @State private var relaySubscriptions = [SubscriptionCancellable]()
-    @State private var performingInitialLoad = true
     @State private var isShowingRelayList = false
-    static let initialLoadTime = 2
+    
+    /// When set to true this will display a fullscreen progress wheel for a set amount of time to give us a chance
+    /// to get some data from relay. The amount of time is defined in `staticLoadTime`.
+    @State private var showTimedLoadingIndicator = true
+    
+    /// The amount of time (in seconds) the loading indicator will be shown when showTimedLoadingIndicator is set to 
+    /// true.
+    static let staticLoadTime: TimeInterval = 2
 
     @ObservedObject var user: Author
 
-    @State private var stories: [Author] = []
-    @State private var selectedStoryAuthor: Author?
-    @State private var storiesCutoffDate = Calendar.current.date(byAdding: .day, value: -2, to: .now)!
+    @State private var showRelayPicker = false
+    @State private var selectedRelay: Relay? 
 
-    private var isShowingStories: Bool {
-        selectedStoryAuthor != nil
-    }
-    
     init(user: Author) {
         self.user = user
-        _authors = FetchRequest(
-            fetchRequest: user.followedWithNewNotes(
-                since: Calendar.current.date(byAdding: .day, value: -2, to: .now)!
-            )
-        )
     }
     
-    /// Downloads the data we need to show stories. 
-    func downloadStories() async {
-        relaySubscriptions.removeAll()
-        
-        let followedKeys = await Array(currentUser.socialGraph.followedKeys)
-        
-        if !followedKeys.isEmpty {
-            let textFilter = Filter(
-                authorKeys: followedKeys, 
-                kinds: [.text, .delete, .repost, .longFormContent, .report], 
-                since: storiesCutoffDate
-            )
-            let textSubs = await relayService.fetchEvents(matching: textFilter)
-            relaySubscriptions.append(textSubs)
+    var homeFeedFetchRequest: NSFetchRequest<Event> {
+        Event.homeFeed(for: user, before: lastRefreshDate, seenOn: selectedRelay) 
+    }
+    
+    var homeFeedFilter: Filter {
+        var filter = Filter(kinds: [.text, .delete, .repost, .longFormContent, .report])
+        if selectedRelay == nil {
+            filter.authorKeys = user.followedKeys.sorted()
+        } 
+        return filter
+    }
+    
+    var navigationBarTitle: LocalizedStringResource {
+        if let relayName = selectedRelay?.host {
+            LocalizedStringResource(stringLiteral: relayName)
+        } else {
+            .localizable.accountsIFollow
         }
     }
-    
+
     var body: some View {
         ZStack {
-            let homeFeedFilter = Filter(
-                authorKeys: user.followedKeys, 
-                kinds: [.text, .delete, .repost, .longFormContent, .report], 
-                limit: 100, 
-                since: nil,
-                keepSubscriptionOpen: true
-            )
             PagedNoteListView(
-                databaseFilter: Event.homeFeed(for: user, before: lastRefreshDate), 
+                databaseFilter: homeFeedFetchRequest,
                 relayFilter: homeFeedFilter,
+                relay: selectedRelay,
                 context: viewContext,
                 tab: .home,
                 header: {
-                    AuthorStoryCarousel(
-                        authors: $stories, 
-                        selectedStoryAuthor: $selectedStoryAuthor
-                    )
+                    EmptyView()
                 },
                 emptyPlaceholder: { _ in
                     VStack {
@@ -87,42 +75,42 @@ struct HomeFeedView: View {
                 },
                 onRefresh: {
                     lastRefreshDate = .now
-                    storiesCutoffDate = Calendar.current.date(byAdding: .day, value: -2, to: lastRefreshDate)!
-                    authors.nsPredicate = user.followedWithNewNotesPredicate(
-                        since: storiesCutoffDate
-                    )
-                    Task { await downloadStories() }
                     return Event.homeFeed(for: user, before: lastRefreshDate)
                 }
             )
             .padding(0)
             
-            StoriesView(
-                cutoffDate: $storiesCutoffDate,
-                authors: stories,
-                selectedAuthor: $selectedStoryAuthor
-            )
-            .scaleEffect(isShowingStories ? 1 : 0.5)
-            .opacity(isShowingStories ? 1 : 0)
-            .animation(.default, value: selectedStoryAuthor)
-            
-            if performingInitialLoad {
+            if showTimedLoadingIndicator {
                 FullscreenProgressView(
-                    isPresented: $performingInitialLoad,
-                    hideAfter: .now() + .seconds(Self.initialLoadTime)
+                    isPresented: $showTimedLoadingIndicator,
+                    hideAfter: .now() + .seconds(Int(Self.staticLoadTime))
                 )
             } 
+            
+            if showRelayPicker {
+                RelayPicker(
+                    selectedRelay: $selectedRelay,
+                    defaultSelection: String(localized: .localizable.accountsIFollow),
+                    author: user,
+                    isPresented: $showRelayPicker
+                )
+                .onChange(of: selectedRelay) { _, _ in
+                    showTimedLoadingIndicator = true
+                    lastRefreshDate = .now + Self.staticLoadTime
+                    Task {
+                        withAnimation {
+                            showRelayPicker = false
+                        }
+                    }
+                }
+            }
         }
         .doubleTapToPop(tab: .home) { _ in
-            if isShowingStories {
-                selectedStoryAuthor = nil
-            } else {
-                NotificationCenter.default.post(
-                    name: .scrollToTop,
-                    object: nil,
-                    userInfo: ["tab": AppDestination.home]
-                )
-            }
+            NotificationCenter.default.post(
+                name: .scrollToTop,
+                object: nil,
+                userInfo: ["tab": AppDestination.home]
+            )
         }
         .background(Color.appBg)
         .toolbar {
@@ -130,54 +118,23 @@ struct HomeFeedView: View {
                 SideMenuButton()
             }
             ToolbarItem(placement: .navigationBarTrailing) {
-                if isShowingStories {
-                    Button {
-                        selectedStoryAuthor = nil
-                    } label: {
-                        Image.stories.rotationEffect(selectedStoryAuthor == nil ? Angle.zero : Angle(degrees: 90))
-                            .animation(.default, value: selectedStoryAuthor)
+                Button {
+                    withAnimation {
+                        showRelayPicker.toggle()
                     }
-                } else {
-                    Button {
-                        isShowingRelayList = true
-                    } label: {
-                        HStack(spacing: 3) {
-                            Image("relay-left")
-                                .colorMultiply(relayService.numberOfConnectedRelays > 0 ? .white : .red)
-                            Text("\(relayService.numberOfConnectedRelays)")
-                                .font(.clarity(.bold, textStyle: .title3))
-                                .foregroundColor(.primaryTxt)
-                            Image("relay-right")
-                                .colorMultiply(relayService.numberOfConnectedRelays > 0 ? .white : .red)
-                        }
-                    }
-                    .sheet(isPresented: $isShowingRelayList) {
-                        NavigationView {
-                            RelayView(author: user)
-                        }
-                    }
+                } label: {
+                    Image(systemName: "line.3.horizontal.decrease.circle")
+                        .foregroundStyle(Color.secondaryTxt)
+                        .accessibilityLabel(Text(.localizable.filter))
                 }
+                .frame(minWidth: 40, minHeight: 40)
             }
         }
         .padding(.top, 1)
-        .nosNavigationBar(title: isShowingStories ? .localizable.stories : .localizable.homeFeed)
-        .task {
-            await downloadStories()
-        }
+        .nosNavigationBar(title: navigationBarTitle)
         .onAppear {
             if router.selectedTab == .home {
                 isVisible = true 
-            }
-            if !isShowingStories {
-                stories = authors.map { $0 }
-            }
-        }
-        .onChange(of: isShowingStories) { _, newValue in
-            if newValue {
-                analytics.enteredStories()
-            } else {
-                analytics.closedStories()
-                stories = authors.map { $0 }
             }
         }
         .onDisappear { isVisible = false }
@@ -192,11 +149,23 @@ struct HomeFeedView: View {
 #Preview {
     var previewData = PreviewData()
     
+    func createTestData() {
+        let user = previewData.alice
+        let addresses = Relay.recommended
+        addresses.forEach { address in
+            let relay = try? Relay.findOrCreate(by: address, context: previewData.previewContext)
+            relay?.relayDescription = "A Nostr relay that aims to cultivate a healthy community."
+            relay?.addToAuthors(user)
+        }
+        
+        _ = previewData.shortNote
+    }
+    
     return NavigationStack {
         HomeFeedView(user: previewData.alice)
     }
     .inject(previewData: previewData)
     .onAppear {
-        _ = previewData.shortNote
+        createTestData()
     }
 }