+++ Nos/Views/Modifiers/LinearGradient+Planetary.swift
@@ -44,12 +44,6 @@ extension LinearGradient {
         endPoint: .bottom
     )
 
-    public static let storiesBackground = LinearGradient(
-        colors: [.storiesBgTop, .storiesBgBottom],
-        startPoint: .top,
-        endPoint: .bottom
-    )
-    
     public static let gold = LinearGradient(
         colors: [Color(hex: "#FFC46B"), Color(hex: "#DE7C21")],
         startPoint: .top,